requirements.txt
google-genai>=1.0.0
sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.2.0
faiss-cpu>=1.8.0
datasets>=2.19.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
scikit-learn>=1.4.0
accelerate>=0.30.0
